Refactor: break out AppRunner and AppRunnerRef to own files
