+++ crates/eframe/src/web/mod.rs
@@ -2,6 +2,8 @@
 
 #![allow(clippy::missing_errors_doc)] // So many `-> Result<_, JsValue>`
 
+mod app_runner;
+mod app_runner_ref;
 pub mod backend;
 mod events;
 mod input;
@@ -11,6 +13,8 @@ pub mod storage;
 mod text_agent;
 mod web_logger;
 
+pub use app_runner::AppRunner;
+pub use app_runner_ref::AppRunnerRef;
 pub use panic_handler::{PanicHandler, PanicSummary};
 pub use web_logger::WebLogger;
 
@@ -33,15 +37,80 @@ pub use backend::*;
 pub use events::*;
 pub use storage::*;
 
-use std::collections::BTreeMap;
-
 use egui::Vec2;
 use wasm_bindgen::prelude::*;
-use web_sys::{EventTarget, MediaQueryList};
+use web_sys::MediaQueryList;
 
 use input::*;
 
-use crate::Theme;
+use crate::{epi, Theme};
+
+// ----------------------------------------------------------------------------
+
+/// Install event listeners to register different input events
+/// and start running the given app.
+///
+/// ``` no_run
+/// #[cfg(target_arch = "wasm32")]
+/// use wasm_bindgen::prelude::*;
+///
+/// /// This is the entry-point for all the web-assembly.
+/// /// This is called from the HTML.
+/// /// It loads the app, installs some callbacks, then returns.
+/// /// It returns a handle to the running app that can be stopped calling `AppRunner::stop_web`.
+/// /// You can add more callbacks like this if you want to call in to your code.
+/// #[cfg(target_arch = "wasm32")]
+/// #[wasm_bindgen]
+/// pub struct WebHandle {
+///     handle: AppRunnerRef,
+/// }
+/// #[cfg(target_arch = "wasm32")]
+/// #[wasm_bindgen]
+/// pub async fn start(canvas_id: &str) -> Result<WebHandle, eframe::wasm_bindgen::JsValue> {
+///     let panic_handler = eframe::web::PanicHandler::install();
+///     let web_options = eframe::WebOptions::default();
+///     eframe::start_web(
+///         canvas_id,
+///         web_options,
+///         Box::new(|cc| Box::new(MyEguiApp::new(cc))),
+///     )
+///     .await
+///     .map(|handle| WebHandle { handle })
+/// }
+/// ```
+///
+/// # Errors
+/// Failing to initialize WebGL graphics.
+pub async fn start_web(
+    canvas_id: &str,
+    panic_handler: PanicHandler,
+    web_options: crate::WebOptions,
+    app_creator: epi::AppCreator,
+) -> Result<AppRunnerRef, JsValue> {
+    #[cfg(not(web_sys_unstable_apis))]
+    log::warn!(
+        "eframe compiled without RUSTFLAGS='--cfg=web_sys_unstable_apis'. Copying text won't work."
+    );
+    let follow_system_theme = web_options.follow_system_theme;
+
+    let mut runner = AppRunner::new(canvas_id, web_options, app_creator).await?;
+    runner.warm_up()?;
+    let runner_ref = AppRunnerRef::new(panic_handler, runner);
+
+    // Install events:
+    {
+        events::install_canvas_events(&runner_ref)?;
+        events::install_document_events(&runner_ref)?;
+        events::install_window_events(&runner_ref)?;
+        text_agent::install_text_agent(&runner_ref)?;
+        if follow_system_theme {
+            events::install_color_scheme_change_event(&runner_ref)?;
+        }
+        events::paint_and_schedule(&runner_ref)?;
+    }
+
+    Ok(runner_ref)
+}
 
 // ----------------------------------------------------------------------------
